+++ packages/tswebextension/CHANGELOG.md
@@ -5,19 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [5.0.3] - 2026-08-25
+
+### Changed
+
+- Updated [@adguard/dnr-converter] to `v1.1.2`.
+- Updated [@adguard/tsurlfilter] to `v6.0.3`.
+- Updated [@adguard/scriptlets] to `v2.5.1`.
+
+[5.0.3]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v5.0.3
+
+## [5.0.2] - 2026-08-24
 
 ### Changed
 
 - Updated [@adguard/dnr-converter] to `v1.1.1`.
 
-[Unreleased]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v5.0.1...HEAD
+[5.0.2]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v5.0.2
 
 ## [5.0.1] - 2026-08-21
+
+### Changed
+
+- Updated [@adguard/agtree] to `v4.2.1`.
 
 ### Fixed
 
 - Short-lived duplicated elements on pages when CSS hits counting is enabled.
+
+[5.0.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v5.0.1
 
 ## [5.0.0] - 2026-07-28
 
@@ -102,7 +118,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [4.1.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v4.1.1
 
-## [v4.1.0-beta.1] - 2026-04-17
+## [4.1.0-beta.1] - 2026-04-17
 
 ### Added
 
@@ -139,7 +155,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `prepare` script from running (`.husky` is excluded via `.dockerignore`).
 - "Block ads manual" doesn't work on tabs opened before the update [AdguardBrowserExtension#3452].
 
-[v4.1.0-beta.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v4.1.0-beta.1
+[4.1.0-beta.1]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v4.1.0-beta.1
 [AdguardBrowserExtension#3393]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3393
 [AdguardBrowserExtension#3329]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3329
 [AdguardBrowserExtension#3414]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3414